% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abre.banco.R
\name{abre.banco}
\alias{abre.banco}
\title{Abrir o banco}
\usage{
abre.banco(banco, itens, bib, disc, disc.cad = 2)
}
\arguments{
\item{banco}{Objeto que tem somente a variável correspondente ao caderno e as variáveis correspondentes às espostas aos itens}

\item{itens}{O(s) objeto(s) com os itens da(s) disciplina(s) que compõe(m) os cadernos no BIB (tem que ser todos, por enquanto; depois tem que melhorar a função)}

\item{bib}{Objeto com o BIB}

\item{disc}{Qual será a disciplina desse banco?}

\item{disc.cad}{Quantidade de disciplinas em cada caderno}
}
\value{
A função retorna uma lista com dois elementos: 'respostas' e 'gabarito'
}
\description{
Construir um banco 'aberto' a partir de uma base com informações de respostas organizada por cadernos distintos.
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
set.seed(1000)
gab.lc = sample (LETTERS[1:4], 9, replace = TRUE)
gab.mt = sample (LETTERS[1:4], 9, replace = TRUE)
itens.lc = data.frame (Bloco = rep (1:3, c (3,3,3)), Posicao = rep (1:3, 3),
                       Item = sample (12345:54321, 9), Origem = 'NOVO',
                       Gabarito = gab.lc, Num_bilog = 201:209, 
                       Nome_bilog = paste ('P', 0001:0009, sep = ''), Disciplina = 'LC')
itens.mt = data.frame (Bloco = rep (1:3, c (3,3,3)), Posicao = rep (1:3, 3),
                       Gabarito = gab.mt, Item = sample (12345:54321, 9),
                       Origem = 'NOVO', Num_bilog = 201:209,
                       Nome_bilog = paste ('P', 0001:0009, sep = ''), Disciplina = 'MT')

bib = data.frame (Caderno = 1:3, Disciplina1 = rep ('LC', 3), Disciplina2 = rep ('MT', 3), 
                  Bloco1 = 1:3, Bloco2 = c(2, 3, 1), Bloco3 = 1:3, Bloco4 = c(2, 3, 1))
itens = rbind (itens.lc, itens.mt)
resp = matrix (sample (LETTERS[1:4], 180, replace = TRUE), ncol = 6)
banco = data.frame (CAD = seq(1:3), resp)
disc = 'LC'
aberto = abre.banco (banco = banco, itens = itens, bib = bib, disc = disc)
}
\author{
Alexandre Jaloto
}
