% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abre.resp.R
\name{abre.resp}
\alias{abre.resp}
\title{Separar o vetor de resposta}
\usage{
abre.resp(unico)
}
\arguments{
\item{unico}{Objeto com o vetor das respostas a todos os itens}
}
\value{
A função retorna uma matriz com número de colunas igual à quantidade de itens e número de linhas igual à quantidade de participantes.
}
\description{
Separa um único vetor de respostas (vários itens) em vetores compostos pela resposta a um único item
}
\details{

}
\examples{
arq.par = 'Rodada6.PAR'
categorias = 5
p = .65
m = 0
dp = 1
int.dp = .25
int.nivel = c (-2, 2)
met = 3
nomes.cat = data.frame (a = paste0 ('a', 1:16), b = paste0 ('b', 1:16),
               c = paste0 ('c', 1:16), d = paste0 ('d', 1:16),
               e = paste0 ('e', 1:16))
aloca.item.multi (arq.par, categorias, p = .65, m = 0,
               dp = 1, int.dp = .25, int.nivel = c (-2, 2),
               met = 3, nomes.cat = nomes.cat)

}
\author{
Alexandre Jaloto
}
