% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeq.par.R
\name{adeq.par}
\alias{adeq.par}
\title{Adequação dos parâmetros estimados}
\usage{
adeq.par(
  arq.par,
  a = list(0.45, 4),
  b = list(-4, 4),
  c = list(0, 0.45),
  erro.a = NULL,
  erro.b = NULL,
  erro.c = NULL,
  salvar = FALSE
)
}
\arguments{
\item{arq.par}{Arquivo PAR geado pelo BILOG-MG}

\item{a}{Vetor com o menor valor aceitável e o maior valor aceitável do parâmetro 'a'}

\item{b}{Vetor com o menor valor aceitável e o maior valor aceitável do parâmetro 'b'}

\item{c}{Vetor com o menor valor aceitável e o maior valor aceitável do parâmetro 'c'}

\item{erro.a}{valor do maior valor aceitável para o erro do parâmetro 'a'}

\item{erro.b}{valor do maior valor aceitável para o erro do parâmetro 'b'}

\item{erro.c}{valor do maior valor aceitável para o erro do parâmetro 'c'}

\item{salvar}{Valor lógico. Indica se o resultado será salvo em um arquivo; caso seja TRUE, o arquivo será salvo como "ITENS_FORA_INTERVALO.txt"}
}
\value{
A função retorna uma data.frame com o número do item no BILOG-MG e o motivo de sua exclusão
}
\description{
Verificar a adequação dos parâmetros dos itens estimados pelo BILOG-MG.
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
adeq.par (arq.par = 'PTLCV3.PAR', a = c (1, 3.5), b = c (-3.5, 3.8), c = c (0, .25))
}
\author{
Alexandre Jaloto
}
