% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloca.dico.R
\name{aloca.dico}
\alias{aloca.dico}
\title{Alocação de itens dicotômicos nos níveis da escala}
\usage{
aloca.dico(
  par,
  met,
  dp = 1,
  int.dp = 0.5,
  int.nivel = c(-2, 2),
  centro = "centro",
  banco,
  min.casos = 50,
  log = TRUE
)
}
\arguments{
\item{par}{Objeto com os parâmetros dos itens estimados. Deve possuir três colunas,
uma para cada parâmetro (nesta ordem: a, b, c).
Só é necessário quando o método for um dos que seguem: \code{b}, \code{b65}, \code{n65} ou \code{(2+c)/3}.}

\item{met}{O método de alocação. As opções são: \code{b} (o parâmetro 'b' é utilizado para
alocar o item no nível); \code{b65} (o valor de theta em que a
probabilidade de acertar o item vale 0.65); \code{b65emp} (o valor de theta em que 65
porcento das pessoas acertaram o item; é a abcissa da
intercessão
da reta 0.65, com a interpolação linear entre a proporção empírica do nível e a do
nível anterior); \code{n65} (o nível para \code{b65}); \code{n65emp}
(o nível em que 65 porcento das pessoas acertaram o item); \code{(2+c)/3} (o valor de theta
em que a probabilidade de acertar o item vale (2+c)/3)}

\item{dp}{Desvio padrão da escala transformada}

\item{int.dp}{Os níveis variam de quanto em quanto, em relação ao \code{dp}?}

\item{int.nivel}{O nível mais baixo e o nível mais alto}

\item{centro}{\code{'centro'} se o nível for centrado no ponto; \code{'esquerda'} se o nível inicia
no ponto (fechado na esquerda)}

\item{banco}{Objeto do tipo \code{data.frame} com as respostas (0 = erro; 1 = acerto) e o
escore de cada indivíduo.
A última variável do banco é oe score. Esse objeto é necessário somente se
\code{met = 'b65emp'} ou \code{met = 'n65emp'}. ATENÇÃO: os nomes das colunas
dos itens devem começar com 'I' e o da nota deve ser 'NOTA'; para detalhes, veja o exemplo.}

\item{min.casos}{Número mínimo de respostas em um determinado nível para que o item
possa ser alocado nele. Usado somente quando o método
for \code{'b65emp'} ou \code{'n65emp'}}

\item{log}{Argumento lóico. A métrica é logística?}
}
\value{

}
\description{
Alocar itens dicotômicos em pontos ou níveis da escala
}
\details{
Para esta função, ainda é necessário pensar sobre as transformações dos
parâmetros para alocação dos itens na escala.
Por exemplo, se de fato vai para escala (0,1) adequadamente e depois para a escala da
avaliação. É possível que haja algum erro
nessas transformações.
}
\examples{
set.seed(1000)
par = as.matrix(data.frame(a = rlnorm (30, 0, .5), b = rnorm (30, 0, 1), c = rbeta (30, 5, 17)))
pop = rnorm (5000, 0, 1)

# precisa ser do tipo data.frame (não pode ser data.table nem matrix)
banco = data.frame(simular (pop, par))

calib = mirt::mirt(banco, 1, itemtype = '3PL', TOL = 0.01)
nota = data.frame(mirt::fscores (calib))
names (nota) = 'NOTA'
par = mirt::coef(calib, simplify = TRUE, IRTpars = TRUE)$items[,1:3]
banco$NOTA = nota

# atenção para a forma de nomear
names (banco) = c(paste0('I', 1:30), 'NOTA')
head (banco)

int.nivel = c(-3, 3)

centro = 'esquerda'
met = 'n65emp'

aloca.dico(par = par, met = met, int.nivel = int.nivel, centro = centro, banco = banco, min.casos = 30)


}
\author{
Alexandre Jaloto
}
