% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloca.pessoa.R
\name{aloca.pessoa}
\alias{aloca.pessoa}
\title{Alocação dos indivíduos nos níveis da escala}
\usage{
aloca.pessoa(escore, niv = seq(-2, 2, 0.25), direita = TRUE)
}
\arguments{
\item{escore}{Vetor com os escores dos indivíduos}

\item{niv}{Os níveis da escala}

\item{direita}{argumento lógico. Se \code{TRUE} (padrão), é fechado na direita e aberto na
esquerda. Se \code{FALSE}, é aberto na direita e fechado na esquerda.}
}
\value{
A função retorna um objeto do tipo lista com dois elementos:

$theta: O escore dos indivíduos

$niveis: O nível de cada indivíduo na escala
}
\description{
Alocar os indivíduos, de acordo com seu escore, nos níveis da escala
}
\details{
O nível é selecionado de acordo com o escore. O participante é posicionado no nível
em que o escore é maior do que o valor mínimo do intervalo e menor ou igual ao maior valor
do intervalo, se \code{direita = TRUE} Se \code{direita = FALSE}, o participante é posicionado no nível
em que o escore é maior ou igual ao valor mínimo do intervalo e menor do que o maior valor
do intervalo.
}
\examples{
set.seed (12345)
escore = rnorm (100)
aloca.pessoa (escore)

}
\author{
Alexandre Jaloto
}
