% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloca.poli.R
\name{aloca.poli}
\alias{aloca.poli}
\title{Alocação de itens politômicos nos níveis da escala}
\usage{
aloca.poli(
  arq.par,
  categorias,
  p = 0.65,
  m = 0,
  dp = 1,
  int.dp = 0.25,
  int.nivel = c(-2, 2),
  met = 3,
  nomes.cat = c()
)
}
\arguments{
\item{arq.par}{Arquivo PAR gerado, pelo MULTILOG, com os parâmetros dos itens estimados}

\item{categorias}{O maior quantitativo de categorias de um mesmo item}

\item{p}{A probabilidade de obter determinado código (categoria) utilizada para
alocar o item em um nível}

\item{m}{Média da escala transformada (o padrão da função é não transformar)}

\item{dp}{Desvio padrão da escala transformada (o padrão da função é não transformar)}

\item{int.dp}{Os níveis variam de quanto em quanto, em relação ao \code{dp}?}

\item{int.nivel}{O nível mais baixo e o nível mais alto}

\item{met}{O método de alocação, podendo ser \code{1}ou \code{3}}

\item{nomes.cat}{objeto do tipo \code{data.frame} ou \code{matrix} indicando o nome
de cada
categoria de cada item (nesta \code{data.frame}, se a categoria não existir em
determinado item, indicar \code{NA})}
}
\value{
A função retorna um objeto do tipo lista com três elementos:

\code{$prof}: O escore que retorna a probabilidade 'p' do
indivíduo selecionar determinada categoria

\code{$niveis}: O nível de cada categoria do item

\code{$MaiorProb}: A categoria mais provável de ser selecionada em cada nível
}
\description{
Alocar os itens com parâmetros estimados pelo MULTILOG nos níveis da escala
}
\details{
Para esta função, ainda é necessário pensar sobre as transformações dos
parâmetros para alocação dos itens na escala.
Por exemplo, se de fato vai para a escala (0,1) adequadamente e depois para a escala da
avaliação. É possível que haja algum erro
nessas transformações.

Se \code{met = 1}, o posicionamento do item é feito da seguinte maneira: calcula-se a
proficiência que o participante deveria ter, dada a probabilidade de 0,65 de ele
marcar aquela categoria ou mais alta

\deqn{\theta = (-1/a)*ln((1-p)/1)+b}

onde \code{a} é a discriminação, \code{p} é a probabilidade de marcar a categoria ou
superior (no caso, 0,65) e \code{b} é o parâmetro de dificuldade.

Se \code{met = 3}, o posicionamento do item é feito da seguinte maneira: calcula-se a
probabilidade de uma categoria ser selecionada dado que o theta é igual ao nível.
Seleciona-se o nível em que há 65\% de probabilidade ou mais de um estudante com
proficiência naquele nível marcar uma categoria ou mais alta.
}
\examples{
arq.par = 'Rodada6.PAR'
categorias = 5
p = .65
m = 0
dp = 1
int.dp = .25
int.nivel = c (-2, 2)
met = 3
nomes.cat = data.frame (a = paste0 ('a', 1:16), b = paste0 ('b', 1:16),
               c = paste0 ('c', 1:16), d = paste0 ('d', 1:16),
               e = paste0 ('e', 1:16))
aloca.poli (arq.par, categorias, p = .65, m = 0,
               dp = 1, int.dp = .25, int.nivel = c (-2, 2),
               met = 3, nomes.cat = nomes.cat)

}
\author{
Alexandre Jaloto
}
