% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arred.R
\name{arred}
\alias{arred}
\title{Arredondamento de números}
\usage{
arred(x, n)
}
\arguments{
\item{x}{Valor a ser arredondado}

\item{n}{Quantidade de casas decimais}
}
\value{
A função retorna o valor arredondado de acordo com o critério determinado.
}
\description{
Arredondar valores numéricos de acordo com a norma 5891:2014 da ABNT. O mesmo critério é adotado por pogramas como Excel e SPSS.
}
\details{
O programa multiplica x por 10^n, trunca esse número e
diminui de x * 10^n. Se essa diferença for maior ou igual a 0.5,
então arredonda para cima; se essa diferença for menor do que 0.5,
então arredonda para baixo.

ATENÇÃO: devido à forma como o número binário é convertido em número real,
é aconselhável utilizar o comando options(digits = 16). O padrão para essa opção é 7.
}
\examples{
x = 100.123456
n = 5
arred (x, n)
options(digits=7)
arred (x, n)
options(digits=16)
arred (x, n)
options(digits=22)
arred (x, n)


}
\author{
Alexandre Jaloto
}
