% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{dif}
\alias{dif}
\title{Análise de DIF}
\usage{
dif(
  arq.exp,
  arq.sco,
  perc = list(5, 95),
  dif.dif = 0.15,
  grupo = 7,
  int.comum = TRUE,
  salvar = FALSE
)
}
\arguments{
\item{arq.exp}{arquivo .EXP gerado pelo BILOG-MG}

\item{arq.sco}{arquivo .SCO gerado pelo BILOG-MG}

\item{perc}{os percentis que definirão o intervalo de análise para detecção de DIF}

\item{dif.dif}{a diferença máxima tolerada entre a proporção de acerto observada e a esperada segundo o modelo}

\item{grupo}{o grupo, no arquivo .BLM, que está em análise; ou seja, o grupo focal corresponde a qual grupo no arquivo .BLM?}

\item{int.comum}{valor lógico: se TRUE, o intervalo a ser analisado é o compreendido entre o maior percentil inferior e o
menor percentil superior.}

\item{salvar}{argumento lógico que indica se a saída será salva em dois arquivos (um para DIF e outro para ajuste do modelo)}
}
\value{
Uma lista com três elementos:

$DIF
Contém os itens que apresentaram DIF entre o grupo focal e algum outro grupo do BILOG-MG.

$INTERVALOS
Os intervalos em que ocorreram as análises. P5 e P95 é o intervalo compreendido entre o maior
P5 e o menor P95 de cada análise entre dois grupos. LIMITE_INF e LIMITE_SUP compreende o
limite em que a análise de fato ocorreu, ou seja, os pontos de quadratura no interior do
intervalo do P\% e P95.

$AJUSTE
contém os itens que apresentaram problemas de ajuste.
}
\description{
Verifica existência de DIF e a qualidade do ajuste de um item calibrado no BILOG-MG
}
\details{
Para analisar DIF, a função diminui a proporção de acerto do grupo focal da proporção de acerto
do grupo de referência. Para verificar a qualidade do ajuste do item, a função diminui a probabilidade de
acerto segundo o modelo da proporção de acerto observada.

Sobre o argumento int.comum: supondo que o intervalo escolhido foi perc = c (.05, .95). A análise de DIF se dará para o intervalo
compreendido entre o maior P5 e o menor P95 dos dois grupos.

O que a função faz?
Ela lê o arquivo expect; seleciona os itens comuns; seleciona o
grupo de interesse e os outros grupos; compara as duas proporções esperadas de resposta correta; seleciona
o grupo de interesse; compara o teórico com a proporção esperada do grupo (para análise de ajuste)

ATENÇÃO: é necessário que o início do arquivo SCO tenha duas linhas que não são referentes a dados
}
\author{
Alexandre Jaloto
}
