% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.nivel.R
\name{freq.nivel}
\alias{freq.nivel}
\title{Frequência dos indivíduos nos níveis
da escala}
\usage{
freq.nivel(escore, peso = 1, niv = seq(-2, 2, 0.25), met = 3)
}
\arguments{
\item{escore}{Vetor com os escores}

\item{peso}{Vetor com o peso de cada escore}

\item{niv}{Os níveis da escala}

\item{met}{O método de alocação, segundo o relatório de 2014 da ANA}
}
\value{
A função retorna um objeto do tipo data.frame com duas variáveis:
o nível da escala e a frequência relativa dos indivíduos.
}
\description{
Verificar a frequência dos indivíduos em cada nível
da escala
}
\details{
Detalhes
}
\examples{
set.seed(1000)
escore = rnorm (100)
freq.nivel (escore)
}
\author{
Alexandre Jaloto
}
