% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera.form.R
\name{gera.form}
\alias{gera.form}
\title{Gera os forms e os gabaritos para o BILOG-MG}
\usage{
gera.form(itens, bib, disc.cad = 2, cad.1 = c(1, 1), carac = 34)
}
\arguments{
\item{itens}{O(s) objeto(s) com os itens da(s) disciplina(s) que compõe(m) os
cadernos no BIB (tem que ser todos, por enquanto; depois tem que melhorar a função).
ATENÇÃO: precisa ser data.frame}

\item{bib}{Objeto com o BIB. ATENÇÃO: precisa ser data.frame}

\item{disc.cad}{Quantidade de disciplinas em cada caderno}

\item{cad.1}{Número do primeiro form, no BILOG-MG, referente aos cadernos de cada disciplina}

\item{carac}{Quantidade de caracteres antes da leitura do vetor de resposta no BILOG-MG}
}
\value{
Para cada disciplina, dois arquivos 'txt':\preformatted{1. arquivo com os forms para o BILOG-MG

2. arquivo com os gabaritos dos forms para o BILOG-MG
}
}
\description{
A partir das informaçõs do BIB e dos itens, esta função gera
dois arquivos 'txt' para serem incorporados aos arquivos do BILOG-MG.
São os forms, que devem ser adicionados no BLM, e os gabaritos.
}
\examples{
set.seed(1000)
gab.lc = sample (LETTERS[1:4], 9, replace = TRUE)
gab.mt = sample (LETTERS[1:4], 9, replace = TRUE)
itens.lc = data.frame (Bloco = rep (1:3, c (3,3,3)), Posicao = rep (1:3, 3),
                       Item = sample (12345:54321, 9), Origem = 'NOVO',
                       Gabarito = gab.lc, Num_bilog = 201:209,
                       Nome_bilog = paste ('P', 0001:0009, sep = ''), Disciplina = 'LC')
itens.mt = data.frame (Bloco = rep (1:3, c (3,3,3)), Posicao = rep (1:3, 3),
                       Gabarito = gab.mt, Item = sample (12345:54321, 9),
                       Origem = 'NOVO', Num_bilog = 201:209,
                       Nome_bilog = paste ('P', 0001:0009, sep = ''), Disciplina = 'MT')

bib = data.frame (Caderno = 1:3, Disciplina1 = rep ('LC', 3), Disciplina2 = rep ('MT', 3),
                  Bloco1 = 1:3, Bloco2 = c(2, 3, 1), Bloco3 = 1:3, Bloco4 = c(2, 3, 1))
itens = rbind (itens.lc, itens.mt)

gera.form (itens, bib, cad.1 = c (12, 14))
}
\author{
Alexandre Jaloto
}
