% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEPsico-package.R
\docType{package}
\name{INEPsico-package}
\alias{INEPsico}
\alias{INEPsico-package}
\title{INEPsico: Bora Facilitar a Vida}
\description{
Automatização de rotinas para análises psicométricas
}
\details{
Este pacote tem como objetivo automatizar algumas rotinas de análise de
itens e testes por meio da Teoria Clássica dos Testes (TCT) e Teoria de Resposta
ao Item (TRI). Além de realizar análises via TCT e produzir relatórios desses
resultados, o pacote cria arquivos para serem lidos no BILOG-MG.
As funções que envolvem análises via TRI são voltadas para
trabalhar com resultados produzidos nesse programa.
}
\author{
Alexandre Jaloto \href{mailto:alexandre.jaloto@inep.gov.br}{alexandre.jaloto@inep.gov.br}
}
