% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.teste.R
\name{info.teste}
\alias{info.teste}
\title{Curva de informação do teste}
\usage{
info.teste(
  pars = list(),
  testes = list(),
  theta = seq(-4, 4, 0.01),
  erro = FALSE,
  ajustar = TRUE
)
}
\arguments{
\item{pars}{Objeto do tipo data.frame, matrix ou uma lista com
data.frame e/ou matrix. Este objeto contém os parâmetros dos
itens de cada instrumento. Se o objetivo é plotar a curva de
apenas um instrumento, o objeto deve ser uma data.frame ou
matrix; se a curva de mais de um instrumento for plotada, o
objeto deve ser uma lista em que cada elemento é uma data.frame
ou matrix com os parâmetros. A primeira coluna do objeto (ou
elemento, em caso de ser uma lista) contém os valores do
parâmetro 'a'; a segunda, os do parâmetro 'b'; a terceira, os
do parâmetro 'c'}

\item{testes}{Vetor com os nomes dos testes, para inserir a
legenda no gráfico}

\item{theta}{Vetor com os valores de traço latente para a
construção do gráfico}

\item{erro}{Valor lógico. Se TRUE, plota também a curva
de erro da medida}

\item{ajustar}{Valor lógico. Se TRUE, plota todas as
curvas de informação na mesma escala. Se FALSE, as curvas
serão plotadas em escalas diferentes no que diz respeito ao
eixo 'y'.}
}
\value{
A função retorna um gráfico com a curva de informação
de cada teste.
}
\description{
Produz a(s) curva(s) de informação do(s) teste(s)
}
\examples{
# criar objeto com os parâmetros
set.seed(1000)
pars1 = data.frame (a = runif (50, .7, 1.3), b = runif (50, -3, 3), c = runif (50, 0, 1))
pars2 = data.frame (a = runif (50, .7, 1.3), b = runif (50, -3, 3), c = runif (50, 0, 1))
pars3 = data.frame (a = runif (50, .7, 1.3), b = runif (50, -3, 3), c = runif (50, 0, 1))

pars = list (pars1, pars2, pars3)
testes = c ('T1', 'T2', 'T3')
theta = seq (-3, 3, .5)

info.teste (pars = pars, theta = theta, testes = testes, erro = FALSE, ajustar = TRUE)

info.teste (pars = pars[[1]], theta = theta, testes = testes[1], erro = TRUE)
}
\author{
Alexandre Jaloto
}
