% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.exp.R
\name{ler.exp}
\alias{ler.exp}
\title{Importação do arquivo EXPECT}
\usage{
ler.exp(arq.exp)
}
\arguments{
\item{arq.exp}{Arquivo .EXP}
}
\value{
A função retorna um objeto do tipo data.frame com os dados do arquivo .EXP
}
\description{
Importar o arquivo .EXP produzido pelo BILOG-MG
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
ler.exp (arq.exp = 'PTLCV3.EXP')
}
\author{
Alexandre Jaloto
}
