% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.par.R
\name{ler.par}
\alias{ler.par}
\title{Importação do arquivo PAR}
\usage{
ler.par(arq.par, prog = "BLM", categorias)
}
\arguments{
\item{arq.par}{Arquivo \code{.PAR}}

\item{prog}{O programa que produziu o arquivo \code{.PAR}. Use \code{'BLM'} (padrão)
para BILOG-MG e \code{'MLM'} para MULTILOG.}

\item{categorias}{Somente se prog = \code{'MLM'}. O número máximo de categorias
de um item.}
}
\value{
A função retorna um objeto do tipo data.frame com os dados do arquivo .PAR
}
\description{
Importar o arquivo \code{.PAR} produzido pelo BILOG-MG ou MULTILOG
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
ler.par ('PTLCV3.PAR')
ler.par ('ANA16.PAR', prog = 'MLM', categorias = 4)
}
\author{
Alexandre Jaloto
}
