% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.sco.R
\name{ler.sco}
\alias{ler.sco}
\title{Importação do arquivo de escore}
\usage{
ler.sco(arq.sco, prog = "BLM")
}
\arguments{
\item{arq.sco}{Arquivo SCO gerado pelo BILOG-MG ou pelo MULTILOG}

\item{prog}{Programa que gerou a base de dados ('BLM' para BILOG-MG e 'MLM' para MULTILOG)}
}
\value{
A função retorna uma data.frame com os escores dos indivíduos
}
\description{
Importar o arquivo de escore do BILOG-MG ou do MULTILOG
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
ler.sco (arq.sco = 'PTLCV3.SCO', prog = 'BLM')
}
\author{
Alexandre Jaloto
}
