% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.exp.R
\name{prop.exp}
\alias{prop.exp}
\title{Proporções esperadas}
\usage{
prop.exp(dados)
}
\arguments{
\item{dados}{objeto gerado pela função ler.exp}
}
\value{
data.frame com o nome do item no BILOG-MG, o grupo em que foi aplicado, a
variável analisada (PROPORÇÃO ou PROPORÇÃO DO MODELO), e a proporção em cada ponto de
quadratura.
}
\description{
"Abre" o objeto gerado pela função ler.exp. Ou seja, verifica as
proporções empíricas e do modelo de acerto do item em cada ponto de quadratura.
}
\author{
Alexandre Jaloto
}
