% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio.tct.R
\name{relatorio.tct}
\alias{relatorio.tct}
\title{Relatório TCT}
\usage{
relatorio.tct(
  disc,
  disc.extenso,
  teste,
  n.itens.comuns,
  n.itens.novos,
  n.alt,
  tct
)
}
\arguments{
\item{disc}{Sigla da disciplina / área}

\item{disc.extenso}{Nome da disciplina / área por extenso}

\item{teste}{Nome do teste}

\item{n.itens.comuns}{Quantidade de itens comuns}

\item{n.itens.novos}{Quantidade de itens novos}

\item{n.alt}{Número máximo de alternativas dos itens}

\item{tct}{Objeto com os resultados da análise clássica (no mesmo padrão do
objeto que a função tct retorna). Importante que o nome das variáveis deste
objeto seja no mesmo padrão do objeto que a função tct retorna}
}
\value{
A função retorna um arquivo HTML com o relatório das análises
psicométricas segundo a Teoria Clássica dos Testes (TCT).
}
\description{
Elabora um relatório com informações das análises psicométricas
segundo a Teoria Clássica dos Testes (TCT)
}
\author{
Alexandre Jaloto
}
