% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular.R
\name{simular}
\alias{simular}
\title{Simular padrões de resposta}
\usage{
simular(theta = seq(-3, 3, by = 0.1), pars, mod = "log")
}
\arguments{
\item{theta}{Vetor com as medidas de traço latente dos indivíduos}

\item{pars}{Objeto do tipo data.frame ou matrix com os parâmetros
dos itens. A primeira coluna deve apresentar os valores do parâmetro
de discrimação; a segunda, os valores de dificuldade (posição);
a terceira, os valores de acerto casual (pseudochute). Para modelos
de dois parâmetros ou um, veja a seção Details.}

\item{mod}{O modelo da distribuição de probabilidade. Use "log"
para logístico e "norm" para normal}
}
\value{
A função retorna um objeto do tipo matrix que contém a
probabilidade de cada indivíduo acertar cada item. As linhas
correspondem aos indivíduos e as colunas, aos itens.
}
\description{
Simular padrões de resposta para itens dicotômicos
segundo a Teoria de Resposta ao Item (modelos logístico e normal).
Verifique se a função simdata do pacote mirt contempla suas necessidades,
pois ela é mais rápida.
}
\details{
A simulação requer itens de três parâmetros. Para realizar
simulação de respostas a itens de dois parâmetros, arbitre o valor
do acerto casual para 0. Para realizar a simulação de respostas a
itens de um parâmetro, considere o acerto casual como 0 e a discriminação
como 1.
}
\examples{
set.seed(1000)
theta = rnorm (50, 0, 1)
pars = matrix ( c (runif (20, .5, 2), rnorm (20, 0, 1),
                   runif (20, .05, .4)), nrow = 20, ncol = 3)
simular (theta = theta, pars = pars)
}
\author{
Alexandre Jaloto
}
